[jit] Move jit testing utils to their own file
This moves `JitTestCase` to its own file so that we can have other jit
test files (ex. `test_jit_py3.py`)

There aren't any code changes, just a move and cleaning up the imports
